+++ Sources/Sentry/Public/SentryOptions.h
@@ -14,6 +14,7 @@ NS_ASSUME_NONNULL_BEGIN
 @class SentryMeasurementValue;
 @class SentryReplayOptions;
 @class SentryScope;
+@class SentryReplayOptions;
 
 NS_SWIFT_NAME(Options)
 @interface SentryOptions : NSObject
@@ -369,6 +370,12 @@ NS_SWIFT_NAME(Options)
 @property (nonatomic, assign) BOOL enablePreWarmedAppStartTracing;
 
 #endif // SENTRY_UIKIT_AVAILABLE
+
+#if SENTRY_TARGET_REPLAY_SUPPORTED
+
+@property (nonatomic, strong) SentryReplayOptions* sessionReplay;
+
+#endif // SENTRY_TARGET_REPLAY_SUPPORTED
 
 /**
  * When enabled, the SDK tracks performance for HTTP requests if auto performance tracking and