+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Features
 
+- Session replay GA ()
 - Show session replay options as replay tags (#4639)
 
 ### Fixes